Fix dialog height type mismatch

+++ NvProfileSwitcher/main.cpp
@@ -1033,7 +1033,7 @@ HWND CreateAppMessageWindow(AppMessageData* data,HWND owner){
         DT_CALCRECT|DT_WORDBREAK|DT_NOPREFIX);
     SelectObject(measureDc,oldFont);
     ReleaseDC(nullptr,measureDc);
-    const int windowHeight=std::max(228,measure.bottom+139);
+    const int windowHeight=std::max(228,static_cast<int>(measure.bottom)+139);
 
     HWND dialog=CreateWindowExW(WS_EX_DLGMODALFRAME|WS_EX_TOPMOST,
         L"NvProfileSwitcherMessage",data->title.c_str(),WS_CAPTION|WS_SYSMENU,
